requirements.txt
# local package
-e .
black

# external requirements
click
coverage
flake8
igdb-api-v4
pre-commit
pytest
python-dotenv>=0.5.1
Sphinx
tox
